+++ src/components/Auth/index.ts
@@ -1,3 +1,6 @@
 import DropdownAuth from "./DropdownAuth";
 import BurgerMenuAuth from "./BurgerMenuAuth";
-export { DropdownAuth, BurgerMenuAuth };
+import LoginByProviders from "./LoginByProviders";
+import LoginFormByEmail from "./LoginFormByEmail";
+
+export { DropdownAuth, BurgerMenuAuth, LoginByProviders, LoginFormByEmail };
